refactor deposit into modular component structure
