+++ frontend/src/components/Common/Deposit/index.js
@@ -1,0 +1,2 @@
+export { default as Deposit } from "./Deposit";
+export { default as DepositSong } from "./parts/DepositSong";
